ajustes nas interfaces 116 e 117 e tela da lista de PDVs

+++ nds-service-client-impl/src/main/java/br/com/abril/nds/service/impl/PoliticaCobrancaServiceImpl.java
@@ -251,7 +251,7 @@ public class PoliticaCobrancaServiceImpl implements PoliticaCobrancaService {
 			
 			if(politica.getFormaCobranca() != null && politica.getFormaCobranca().getFatorVencimento() != null) {
 				
-				parametroCobrancaDTO.setFatorVencimento(politica.getFormaCobranca().getFatorVencimento().longValue());
+				parametroCobrancaDTO.setFatorVencimento(politica.getFormaCobranca().getFatorVencimento());
 				
 			}
